requirements.txt
pwntools
pycryptodome
pillow
pyzbar